requirements.txt
amqp==5.0.1
attrs==19.3.0
Automat==0.8.0
billiard==3.6.3.0
blinker==1.4
celery==5.0.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
cloud-init==20.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
Flask==1.1.1
gunicorn==20.0.4
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==16.10.1
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
jq==1.1.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.0.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==1.1.0
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
py==1.9.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
pytest==6.1.1
python-apt==2.0.0+ubuntu0.20.4.1
python-debian===0.1.36ubuntu1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
ssh-import-id==5.10
stopit==1.1.2
systemd-python==234
timeout-decorator==0.4.1
toml==0.10.1
Twisted==18.9.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
vine==5.0.0
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==0.16.1
zipp==1.0.0
zope.interface==4.7.1
